Fix sharemodal flicker
